Hex Editor - convert addresses from int to long, fixes Go to address on 32bit domains, highlighting on addresses after 7FFFFFFF is still busted though

+++ BizHawk.Client.EmuHawk/tools/ToolHelpers.cs
@@ -210,7 +210,7 @@ namespace BizHawk.Client.EmuHawk
 			}
 		}
 
-		public static void ViewInHexEditor(MemoryDomain domain, IEnumerable<int> addresses, Watch.WatchSize size)
+		public static void ViewInHexEditor(MemoryDomain domain, IEnumerable<long> addresses, Watch.WatchSize size)
 		{
 			GlobalWin.Tools.Load<HexEditor>();
 			GlobalWin.Tools.HexEditor.SetToAddresses(addresses, domain, size);
